+++ src/lib.rs
@@ -239,7 +239,8 @@ pub fn parse_iter(text: &str) -> ParseIter<'_, 'static> {
 }
 
 /// A release note for a version.
-#[derive(Debug, Clone)]
+#[allow(single_use_lifetimes)] // https://github.com/rust-lang/rust/issues/69952
+#[derive(Debug, Clone, PartialEq, Eq)]
 #[non_exhaustive]
 pub struct Release<'a> {
     /// The version of this release.