+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - [Add `parse_iter` for partial parsing changelog.](https://github.com/taiki-e/parse-changelog/pull/9)
 
+- Implement `PartialEq` and `Eq` for `Release`.
+
 - [Fix bug in parsing of atx style heading.](https://github.com/taiki-e/parse-changelog/pull/9)
 
 ## [0.3.0] - 2021-04-12